requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
chardet==5.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
django-cors-headers==3.14.0
django-mptt==0.14.0
drf-yasg==1.21.5
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonmerge==1.9.0
jsonschema==4.17.3
MarkupSafe==2.1.2
Naked==0.1.32
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
psycopg2-binary==2.9.5
pycryptodome==3.17
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytz==2022.7.1
PyJWE==1.0.0
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sendgrid==6.9.7
shellescape==3.8.1
simplejson==3.18.4
six==1.16.0
sqlparse==0.4.3
swagger-spec-validator==3.0.3
twilio==7.16.5
uritemplate==4.1.1
urllib3==1.26.15
usaepay==2.1.6
xlrd==2.0.1
zipp==3.15.0
WeasyPrint==52.5